refactor(viewer): move consts to annotation
