+++ libs/annotation/src/const.ts
@@ -1,1 +1,21 @@
+export const CIRCLE_EDIT_POINT_ANGLE = { START_ANGLE: -45, END_ANGLE: 135 }
+
+export const CIRCLE_TEXT_POINT_ANGLE = 135
+
+export const CIRCLE_TEXT_POSITION_SPACING = {
+  x: 75,
+  y: 10,
+}
+export const EDIT_CIRCLE_RADIUS = 3
+export const HALF_OF_RULER_TEXT_BOX = 4
+
+export const LINE_TEXT_POSITION_SPACING = {
+  x: 2,
+  y: 5,
+}
+
+export const RULER_TEXT_POSITION_SPACING = 40
+
+export const TEXT_PADDING = 4
+
 export const IMAGER_PIXEL_SPACING = 'x00181164'
