+++ libs/insight-viewer/src/const/index.ts
@@ -48,28 +48,6 @@ export const ERROR_MESSAGE = {
   ENABLED_ELEMENT_NOT_READY: 'enabledElement value is null, Please check the enabledElement value.',
 } as const
 
-export const LINE_TEXT_POSITION_SPACING = {
-  x: 2,
-  y: 5,
-}
-
-export const RULER_TEXT_POSITION_SPACING = 40
-
-export const HALF_OF_RULER_TEXT_BOX = 4
-
-export const CIRCLE_TEXT_POSITION_SPACING = {
-  x: 75,
-  y: 10,
-}
-
-export const EDIT_CIRCLE_RADIUS = 3
-
-export const CIRCLE_EDIT_POINT_ANGLE = { START_ANGLE: -45, END_ANGLE: 135 }
-
-export const CIRCLE_TEXT_POINT_ANGLE = 135
-
-export const TEXT_PADDING = 4
-
 /**
  * The Annotation Overlay z index should keep the highest value.
  * When using z index, please use a value below that value.
